social-engineering-analyzer/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
qdrant-client>=1.7.1
sentence-transformers>=2.2.0
web3>=6.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0


